% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-eo-list.R
\name{app_eo_list}
\alias{app_eo_list}
\title{Retieve a data frame of metadata about Executive Orders issues in a given year}
\usage{
app_eo_list(year)
}
\arguments{
\item{year}{a value > 1826 and less then last month's year}
}
\value{
data frame (tibble) of \code{potus}, \code{eo_date}, \code{eo_title} and \code{eo_id};
Use \code{eo_id} for calls to \code{\link[=app_get_eo]{app_get_eo()}}.
}
\description{
Retieve a data frame of metadata about Executive Orders issues in a given year
}
\examples{
app_eo_list(1826)
app_eo_list(2014)
}
\references{
\url{http://www.presidency.ucsb.edu/data/orders.php}
}
