% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-get-eo.R
\name{app_get_eo}
\alias{app_get_eo}
\title{Retrieve metadata and full text of an executive order by APP id}
\usage{
app_get_eo(eo_id)
}
\arguments{
\item{eo_id}{a valid APP EO id (via \code{\link[=app_eo_list]{app_eo_list()}})}
}
\value{
data frame (tibble) with \code{date}, \code{eo_num}, \code{title}, \code{potus}, \code{eo_text};
NOTE that \code{eo_num} will be \code{NA} for older EOs without actual numbers.
The \code{title} and \code{potus} and \code{date} should match up with the metadata
retrieved in \code{\link[=app_eo_list]{app_eo_list()}}. Fields may be \code{NA} also due them being
missing.
}
\description{
The APP has an id for each executive order. Use \code{\link[=app_eo_list]{app_eo_list()}} to
find them and then call this function with one of those \code{eo_id}s.
}
\examples{
app_get_eo(104775)
}
\references{
\url{http://www.presidency.ucsb.edu/data/orders.php}
}
