% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-get-ia.R
\name{app_get_inaugurals}
\alias{app_get_inaugurals}
\title{Retrieve metadata and full text for Presidential Inaugural Addresses}
\usage{
app_get_inaugurals()
}
\description{
Retrieve metadata and full text for Presidential Inaugural Addresses
}
\examples{
inaug_df <- app_get_inaugurals()
}
\references{
\url{http://www.presidency.ucsb.edu/inaugurals.php}
}
