% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpresidente-package.R
\docType{package}
\name{elpresidente}
\alias{elpresidente}
\alias{elpresidente-package}
\title{Search and Extract Corpus Elements from 'The American Presidency Project'}
\description{
Search and Extract Corpus Elements from 'The American Presidency Project'
}
\author{
Bob Rudis (bob@rud.is)
}
